+++ fuzz/run.sh
@@ -54,7 +54,7 @@ SECONDS_PER_TARGET="${AC3FORGE_FUZZ_SECONDS:-60}"
 # of the default list - see target_list.
 readonly BASE_TARGETS=(fuzz_scan fuzz_ac3_decode fuzz_eac3_decode fuzz_wav_read
                        fuzz_iec61937_unwrap fuzz_emdf_parse fuzz_oamd_parse
-                       fuzz_joc_parse fuzz_signing_verify fuzz_matroska_demux
+                       fuzz_joc_parse fuzz_osc_parse fuzz_signing_verify fuzz_matroska_demux
                        fuzz_mp4_demux fuzz_mpegts_demux)
 
 adm_enabled() { [ -n "${AC3FORGE_FUZZ_ADM:-}" ]; }